add youtuber detail page rest complete

See merge request webmobile2-sub2/s02p12a108!50

+++ spring-skeleton/src/main/java/com/ssafy/model/dto/Trend.java
@@ -8,7 +8,7 @@ public class Trend {
 	private Date recordDate;
 	private int pointSubscriber;
 	private int difSubscriber;
-	private int pointView;
+	private long pointView;
 	private int difView;
 
 	public int getTno() {
@@ -41,10 +41,10 @@ public class Trend {
 	public void setDifSubscriber(int difSubscriber) {
 		this.difSubscriber = difSubscriber;
 	}
-	public int getPointView() {
+	public long getPointView() {
 		return pointView;
 	}
-	public void setPointView(int pointView) {
+	public void setPointView(long pointView) {
 		this.pointView = pointView;
 	}
 	public int getDifView() {
